use new layout for video
