+++ enjoy/src/i18n/en.json
@@ -268,8 +268,7 @@
   "editResource": "edit resource",
   "deleteResource": "delete resource",
   "deleteResourceConfirmation": "Are you sure to delete {{name}}?",
-  "transcribeAudioConfirmation": "It will remove the old transcription. Are you sure to transcribe {{name}}",
-  "transcribeVideoConfirmation": "It will remove the old transcription. Are you sure to transcribe {{name}}",
+  "transcribeMediaConfirmation": "It will remove the old transcription. Are you sure to transcribe {{name}}",
   "localFile": "local file",
   "resourcesYouAddedRecently": "resources you added recently",
   "recentlyAdded": "recently added",
