--- enjoy/src/renderer/components/audios/audio-transcription.tsx
@@ -1,125 +1,0 @@
-import { useEffect, useContext, useRef } from "react";
-import { MediaPlayerProviderContext } from "@renderer/context";
-import { t } from "i18next";
-import {
-  Button,
-  AlertDialog,
-  AlertDialogTrigger,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogContent,
-  AlertDialogTitle,
-  AlertDialogDescription,
-  AlertDialogCancel,
-  AlertDialogAction,
-  PingPoint,
-  ScrollArea,
-} from "@renderer/components/ui";
-import { LoaderIcon, CheckCircleIcon, MicIcon } from "lucide-react";
-
-export const AudioTranscription = () => {
-  const containerRef = useRef<HTMLDivElement>();
-  const {
-    media,
-    currentSegmentIndex,
-    wavesurfer,
-    setCurrentSegmentIndex,
-    transcription,
-    generateTranscription,
-    transcribing,
-    transcribingProgress,
-  } = useContext(MediaPlayerProviderContext);
-
-  useEffect(() => {
-    if (!containerRef?.current) return;
-
-    containerRef.current
-      ?.querySelector(`#segment-${currentSegmentIndex}`)
-      ?.scrollIntoView({
-        block: "center",
-        inline: "center",
-      } as ScrollIntoViewOptions);
-  }, [currentSegmentIndex, transcription, containerRef]);
-
-  if (!transcription?.result) {
-    return null;
-  }
-
-  return (
-    <ScrollArea
-      ref={containerRef}
-      className="border rounded-lg shadow-lg"
-      data-testid="media-transcription-result"
-    >
-      <div className="sticky top-0 px-4 py-2 bg-background z-10">
-        <div className="flex items-cener justify-between">
-          <div className="flex items-center space-x-2">
-            {transcribing || transcription.state === "processing" ? (
-              <>
-                <PingPoint colorClassName="bg-yellow-500" />
-                <div className="text-sm">
-                  {transcribingProgress > 0 && `${transcribingProgress}%`}
-                </div>
-              </>
-            ) : transcription.state === "finished" ? (
-              <CheckCircleIcon className="text-green-500 w-4 h-4" />
-            ) : (
-              <PingPoint colorClassName="bg-mute" />
-            )}
-            <span className="capitalize">{t("transcript")}</span>
-          </div>
-          <AlertDialog>
-            <AlertDialogTrigger asChild>
-              <Button
-                variant="outline"
-                disabled={transcribing || transcription.state === "processing"}
-                className="capitalize"
-              >
-                {(transcribing || transcription.state === "processing") && (
-                  <LoaderIcon className="animate-spin w-4 mr-2" />
-                )}
-                {transcription.result ? t("regenerate") : t("transcribe")}
-              </Button>
-            </AlertDialogTrigger>
-            <AlertDialogContent>
-              <AlertDialogHeader>
-                <AlertDialogTitle>{t("transcribe")}</AlertDialogTitle>
-                <AlertDialogDescription>
-                  {t("transcribeAudioConfirmation", {
-                    name: media.name,
-                  })}
-                </AlertDialogDescription>
-              </AlertDialogHeader>
-              <AlertDialogFooter>
-                <AlertDialogCancel>{t("cancel")}</AlertDialogCancel>
-                <AlertDialogAction onClick={generateTranscription}>
-                  {t("transcribe")}
-                </AlertDialogAction>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialog>
-        </div>
-      </div>
-
-      {transcription.result.map((t, index) => (
-        <div
-          key={index}
-          id={`segment-${index}`}
-          className={`py-2 px-4 cursor-pointer hover:bg-yellow-400/10 ${
-            currentSegmentIndex === index ? "bg-yellow-400/25" : ""
-          }`}
-          onClick={() => {
-            wavesurfer.seekTo(t.offsets.from / 1000.0 / media.duration);
-            wavesurfer.setScrollTime(t.offsets.from / 1000.0);
-            setCurrentSegmentIndex(index);
-          }}
-        >
-          <div className="flex items-center justify-between">
-            <span className="text-xs opacity-50">#{index + 1}</span>
-          </div>
-          <p className="">{t.text}</p>
-        </div>
-      ))}
-    </ScrollArea>
-  );
-};
