+++ enjoy/src/renderer/components/audios/index.ts
@@ -1,6 +1,5 @@
 export * from "./audios-table";
 export * from "./audio-edit-form";
-export * from "./audio-detail";
 
 export * from "./audios-component";
 export * from "./audible-books-segment";
@@ -8,10 +7,3 @@ export * from "./audios-segment";
 export * from "./audio-card";
 
 export * from "./audio-player";
-export * from "./audio-player-controls";
-export * from "./audio-caption";
-export * from "./audio-info-panel";
-export * from "./audio-recordings";
-export * from "./audio-current-recording";
-export * from "./audio-recorder";
-export * from "./audio-transcription";
