+++ enjoy/src/renderer/components/medias/index.ts
@@ -1,6 +1,9 @@
-export * from "./add-media-button";
-export * from "./media-player";
 export * from "./media-player-controls";
 export * from "./media-caption";
+export * from "./media-info-panel";
+export * from "./media-recordings";
+export * from "./media-current-recording";
+export * from "./media-recorder";
 export * from "./media-transcription";
 export * from "./media-loading-modal";
+export * from "./add-media-button";
