+++ enjoy/src/renderer/components/medias/media-transcription.tsx
@@ -1,4 +1,8 @@
+import { useEffect, useContext, useRef } from "react";
+import { MediaPlayerProviderContext } from "@renderer/context";
+import { t } from "i18next";
 import {
+  Button,
   AlertDialog,
   AlertDialogTrigger,
   AlertDialogFooter,
@@ -8,182 +12,114 @@ import {
   AlertDialogDescription,
   AlertDialogCancel,
   AlertDialogAction,
-  Skeleton,
-  ScrollArea,
-  Button,
   PingPoint,
+  ScrollArea,
 } from "@renderer/components/ui";
-import React, { useEffect, useContext, useState } from "react";
-import { t } from "i18next";
 import { LoaderIcon, CheckCircleIcon, MicIcon } from "lucide-react";
-import {
-  DbProviderContext,
-  AppSettingsProviderContext,
-  AISettingsProviderContext,
-} from "@renderer/context";
 
-export const MediaTranscription = (props: {
-  transcription: TranscriptionType;
-  progress: number;
-  transcribe: () => void;
-  transcribing: boolean;
-  mediaId: string;
-  mediaType: "Audio" | "Video";
-  mediaName?: string;
-  currentSegmentIndex?: number;
-  onSelectSegment?: (index: number) => void;
-}) => {
-  const { addDblistener, removeDbListener } = useContext(DbProviderContext);
-  const { whisperConfig } = useContext(AISettingsProviderContext);
-  const { EnjoyApp } = useContext(AppSettingsProviderContext);
+export const MediaTranscription = () => {
+  const containerRef = useRef<HTMLDivElement>();
   const {
-    transcription,
-    transcribing,
-    progress,
-    transcribe,
-    mediaId,
-    mediaType,
-    mediaName,
+    media,
     currentSegmentIndex,
-    onSelectSegment,
-  } = props;
-  const containerRef = React.createRef<HTMLDivElement>();
-
-  const [recordingStats, setRecordingStats] =
-    useState<SegementRecordingStatsType>([]);
-
-  const fetchSegmentStats = async () => {
-    if (!mediaId) return;
-
-    EnjoyApp.recordings.groupBySegment(mediaId, mediaType).then((stats) => {
-      setRecordingStats(stats);
-    });
-  };
+    wavesurfer,
+    setCurrentSegmentIndex,
+    transcription,
+    generateTranscription,
+    transcribing,
+    transcribingProgress,
+  } = useContext(MediaPlayerProviderContext);
 
   useEffect(() => {
-    addDblistener(fetchSegmentStats);
-    fetchSegmentStats();
+    if (!containerRef?.current) return;
 
-    return () => {
-      removeDbListener(fetchSegmentStats);
-    };
-  }, [transcription]);
-
-  useEffect(() => {
     containerRef.current
       ?.querySelector(`#segment-${currentSegmentIndex}`)
       ?.scrollIntoView({
         block: "center",
         inline: "center",
       } as ScrollIntoViewOptions);
-  }, [currentSegmentIndex, transcription]);
+  }, [currentSegmentIndex, transcription, containerRef]);
 
-  if (!transcription)
-    return (
-      <div className="p-4 w-full">
-        <TranscriptionPlaceholder />
-      </div>
-    );
+  if (!transcription?.result) {
+    return null;
+  }
 
   return (
-    <div
-      className="w-full h-full flex flex-col"
-      data-testid="media-transcription"
+    <ScrollArea
+      ref={containerRef}
+      className="border rounded-lg shadow-lg"
+      data-testid="media-transcription-result"
     >
-      <div className="mb-4 flex items-cener justify-between">
-        <div className="flex items-center space-x-2">
-          {transcribing || transcription.state === "processing" ? (
-            <>
-              <PingPoint colorClassName="bg-yellow-500" />
-              <div className="text-sm">
-                {whisperConfig.service === "local" && `${progress}%`}
-              </div>
-            </>
-          ) : transcription.state === "finished" ? (
-            <CheckCircleIcon className="text-green-500 w-4 h-4" />
-          ) : (
-            <PingPoint colorClassName="bg-mute" />
-          )}
-          <span className="capitalize">{t("transcript")}</span>
+      <div className="sticky top-0 px-4 py-2 bg-background z-10">
+        <div className="flex items-cener justify-between">
+          <div className="flex items-center space-x-2">
+            {transcribing || transcription.state === "processing" ? (
+              <>
+                <PingPoint colorClassName="bg-yellow-500" />
+                <div className="text-sm">
+                  {transcribingProgress > 0 && `${transcribingProgress}%`}
+                </div>
+              </>
+            ) : transcription.state === "finished" ? (
+              <CheckCircleIcon className="text-green-500 w-4 h-4" />
+            ) : (
+              <PingPoint colorClassName="bg-mute" />
+            )}
+            <span className="capitalize">{t("transcript")}</span>
+          </div>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <Button
+                variant="outline"
+                disabled={transcribing || transcription.state === "processing"}
+                className="capitalize"
+              >
+                {(transcribing || transcription.state === "processing") && (
+                  <LoaderIcon className="animate-spin w-4 mr-2" />
+                )}
+                {transcription.result ? t("regenerate") : t("transcribe")}
+              </Button>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>{t("transcribe")}</AlertDialogTitle>
+                <AlertDialogDescription>
+                  {t("transcribeMediaConfirmation", {
+                    name: media.name,
+                  })}
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel>{t("cancel")}</AlertDialogCancel>
+                <AlertDialogAction onClick={generateTranscription}>
+                  {t("transcribe")}
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
-        <AlertDialog>
-          <AlertDialogTrigger asChild>
-            <Button
-              disabled={transcribing || transcription.state === "processing"}
-              className="capitalize"
-            >
-              {(transcribing || transcription.state === "processing") && (
-                <LoaderIcon className="animate-spin w-4 mr-2" />
-              )}
-              {transcription.result ? t("regenerate") : t("transcribe")}
-            </Button>
-          </AlertDialogTrigger>
-          <AlertDialogContent>
-            <AlertDialogHeader>
-              <AlertDialogTitle>{t("transcribe")}</AlertDialogTitle>
-              <AlertDialogDescription>
-                {t("transcribeAudioConfirmation", {
-                  name: mediaName,
-                })}
-              </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-              <AlertDialogCancel>{t("cancel")}</AlertDialogCancel>
-              <AlertDialogAction onClick={transcribe}>
-                {t("transcribe")}
-              </AlertDialogAction>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialog>
       </div>
 
-      {transcription?.result ? (
-        <ScrollArea
-          ref={containerRef}
-          className="flex-1 px-2"
-          data-testid="media-transcription-result"
+      {transcription.result.map((t, index) => (
+        <div
+          key={index}
+          id={`segment-${index}`}
+          className={`py-2 px-4 cursor-pointer hover:bg-yellow-400/10 ${
+            currentSegmentIndex === index ? "bg-yellow-400/25" : ""
+          }`}
+          onClick={() => {
+            wavesurfer.seekTo(t.offsets.from / 1000.0 / media.duration);
+            wavesurfer.setScrollTime(t.offsets.from / 1000.0);
+            setCurrentSegmentIndex(index);
+          }}
         >
-          {transcription.result.map((t, index) => (
-            <div
-              key={index}
-              id={`segment-${index}`}
-              className={`py-1 px-2 mb-2 cursor-pointer hover:bg-yellow-400/25 ${
-                currentSegmentIndex === index ? "bg-yellow-400/25" : ""
-              }`}
-              onClick={() => {
-                onSelectSegment?.(index);
-              }}
-            >
-              <div className="flex items-center justify-between">
-                <span className="text-xs opacity-50">#{index + 1}</span>
-
-                <div className="flex items-center space-x-2">
-                  {(recordingStats || []).findIndex(
-                    (s) => s.referenceId === index
-                  ) !== -1 && <MicIcon className="w-3 h-3 text-sky-500" />}
-                  <span className="text-xs opacity-50">
-                    {t.timestamps.from.split(",")[0]}
-                  </span>
-                </div>
-              </div>
-              <p className="">{t.text}</p>
-            </div>
-          ))}
-        </ScrollArea>
-      ) : (
-        <TranscriptionPlaceholder />
-      )}
-    </div>
-  );
-};
-
-export const TranscriptionPlaceholder = () => {
-  return (
-    <div className="p-4">
-      {Array.from({ length: 5 }).map((_, i) => (
-        <Skeleton key={i} className="h-4 w-full mb-4" />
+          <div className="flex items-center justify-between">
+            <span className="text-xs opacity-50">#{index + 1}</span>
+          </div>
+          <p className="">{t.text}</p>
+        </div>
       ))}
-      <Skeleton className="h-4 w-3/5" />
-    </div>
+    </ScrollArea>
   );
 };
